+++ rustler_tests/native/rustler_serde_test/src/test.rs
@@ -33,19 +33,19 @@ pub fn test<'a>(
 
         // Signed Integers
         "i8 (min)" => run_test!(i8::min_value()),
-        "i8 (0)" => run_test!(0 as i8),
+        "i8 (0)" => run_test!(0_i8),
         "i8 (max)" => run_test!(i8::max_value()),
         "i16 (min)" => run_test!(i16::min_value()),
-        "i16 (0)" => run_test!(0 as i16),
+        "i16 (0)" => run_test!(0_i16),
         "i16 (max)" => run_test!(i16::max_value()),
         "i32 (min)" => run_test!(i32::min_value()),
-        "i32 (0)" => run_test!(0 as i32),
+        "i32 (0)" => run_test!(0_i32),
         "i32 (max)" => run_test!(i32::max_value()),
         "i64 (min)" => run_test!(i64::min_value()),
-        "i64 (0)" => run_test!(0 as i64),
+        "i64 (0)" => run_test!(0_i64),
         "i64 (max)" => run_test!(i64::max_value()),
         "i128 (min)" => run_test!(i128::min_value()),
-        "i128 (0)" => run_test!(0 as i128),
+        "i128 (0)" => run_test!(0_i128),
         "i128 (max)" => run_test!(i128::max_value()),
 
         // Unsigned Integers
